[Resource][Locale] Move LocaleProviderInterface to Resource
