+++ src/Sylius/Bundle/LocaleBundle/Listener/RequestLocaleSetter.php
@@ -4,7 +4,7 @@ namespace Sylius\Bundle\LocaleBundle\Listener;
 
 use Sylius\Component\Locale\Context\LocaleContextInterface;
 use Sylius\Component\Locale\Context\LocaleNotFoundException;
-use Sylius\Component\Locale\Provider\LocaleProviderInterface;
+use Sylius\Component\Resource\Provider\LocaleProviderInterface;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 
 /**
