+++ src/Sylius/Bundle/LocaleBundle/spec/Listener/RequestLocaleSetterSpec.php
@@ -7,7 +7,7 @@ use Prophecy\Argument;
 use Sylius\Bundle\LocaleBundle\Listener\RequestLocaleSetter;
 use Sylius\Component\Locale\Context\LocaleContextInterface;
 use Sylius\Component\Locale\Context\LocaleNotFoundException;
-use Sylius\Component\Locale\Provider\LocaleProviderInterface;
+use Sylius\Component\Resource\Provider\LocaleProviderInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
