+++ src/Sylius/Bundle/ResourceBundle/test/app/AppKernel.php
@@ -33,9 +33,7 @@ class AppKernel extends Kernel
             new Doctrine\Bundle\DoctrineBundle\DoctrineBundle(),
             new Symfony\Bundle\TwigBundle\TwigBundle(),
             new \winzou\Bundle\StateMachineBundle\winzouStateMachineBundle(),
-            new Sensio\Bundle\GeneratorBundle\SensioGeneratorBundle(),
             new AppBundle\AppBundle(),
-
         ];
     }
