+++ src/Sylius/Bundle/ShopBundle/Controller/LocaleSwitchController.php
@@ -13,7 +13,7 @@ namespace Sylius\Bundle\ShopBundle\Controller;
 
 use Sylius\Component\Core\Locale\Handler\LocaleChangeHandlerInterface;
 use Sylius\Component\Locale\Context\LocaleContextInterface;
-use Sylius\Component\Locale\Provider\LocaleProviderInterface;
+use Sylius\Component\Resource\Provider\LocaleProviderInterface;
 use Symfony\Bundle\FrameworkBundle\Templating\EngineInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
