+++ src/Sylius/Component/Locale/spec/Provider/LocaleProviderSpec.php
@@ -14,7 +14,7 @@ namespace spec\Sylius\Component\Locale\Provider;
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Locale\Model\LocaleInterface;
 use Sylius\Component\Locale\Provider\LocaleProvider;
-use Sylius\Component\Locale\Provider\LocaleProviderInterface;
+use Sylius\Component\Resource\Provider\LocaleProviderInterface;
 use Sylius\Component\Resource\Repository\RepositoryInterface;
 
 /**
