+++ src/Sylius/Component/Resource/Provider/LocaleProviderInterface.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Component\Locale\Provider;
+namespace Sylius\Component\Resource\Provider;
 
 /**
  * @author Paweł Jędrzejewski <pawel@sylius.org>
